+++ bundle-switching-sdk/src/service/Services.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Net.Http;
 using System.Threading.Tasks;
-using Coin.Sdk.BS.Messages.V4;
+using Coin.Sdk.BS.Messages.V5;
 
 namespace Coin.Sdk.BS.Service
 {