+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Version 1.4.0
+
+Added:
+
+- Support for Bundle Switching v5; removes the business field from contract termination requests
+and contract termination request answers
+
 ## Version 1.3.0
 
 Added: